# src/docs/swagger/ApiDoc.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Task Manager API",
    "version": "0.1.0",
    "description": "RESTful API For Task Manager App"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "idJWT": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "tags": [
    {
      "name": "User",
      "description": "User APIs"
    }
  ],
  "paths": {
    "/user/create-user": {
      "post": {
        "security": [
          {
            "idJWT": []
          }
        ],
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add new user to database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "flag": {
                      "type": "boolean",
                      "description": "request fail or not"
                    },
                    "data": {
                      "type": "object"
                    },
                    "message": {
                      "type": "string",
                      "description": "message to request"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return JWT if login success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "flag": {
                      "type": "boolean",
                      "description": "request fail or not"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "description": "jwt token"
                        },
                        "exp": {
                          "type": "integer",
                          "description": "token expire date by seconds"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "description": "message to request"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}